refactor(graph): extract StorageMixin and base helper classes (PR 4)
